livekit-agent/requirements.txt
livekit-agents[openai]>=0.11.0
livekit>=0.17.0
python-dotenv
aiohttp>=3.9.0
amadeus>=12.0.0